refactor: extract SubCaseCompletionService from listener (engine#252)
SubCaseCompletionListener was carrying all business logic alongside its
CDI event-delivery responsibility. Extracted all logic into a new
SubCaseCompletionService; listener is now a thin one-line delegate.

Tests that bypass @ObservesAsync by calling onCaseLifecycle directly
continue to work unchanged — the listener still exposes that public
method, now via delegation.

Closes #252
